- Massive progress on scripting.  - It is now possible to attach a PhysicsObject to a CoffeeInstanceData* object and make it move and rotate accordingly, all configured from QtScript. (Even constructing the PhysicsObject)  - Still need to make classes more script-friendly by providing VectorValue objects  - Input can be configured from QtScript now, but requires VectorValue objects to work better.  - Fixed some flaws here and there  - New idea for shader configuration, binding it according to instructions from QtScript. This will use ShaderVariant heavily to provide std::function objects.

+++ general/shadervariant.cpp
@@ -32,9 +32,9 @@ ShaderVariant::~ShaderVariant()
 {
     delete data;
 }
-ShaderVariant::ShaderVariant(std::function<double ()> func)
+ShaderVariant::ShaderVariant(std::function<float ()> func)
 {
-    data = new MutableDataContainer<std::function<double()>>(this,func);
+    data = new MutableDataContainer<std::function<float()>>(this,func);
     type = ShaderDub;
 }
 
@@ -74,7 +74,7 @@ std::function<glm::mat4 ()> ShaderVariant::getMat4() const
         return ((MutableDataContainer<std::function<glm::mat4 ()>>*)data)->getValue();
     return nullptr;
 }
-std::function<double ()> ShaderVariant::getDouble() const
+std::function<float ()> ShaderVariant::getDouble() const
 {
     if(type==ShaderDub)
         return ((MutableDataContainer<std::function<double()>>*)data)->getValue();
